moved member models to models folder
